1.- Changed Murmur3 algo to match origin implementation. 2.- Added three more testing files
