+++ Split/SplitEngine/Hashing/Murmur3Hash.swift
@@ -151,7 +151,7 @@ public final class Murmur3Hash {
             word |= UInt32(bytes[i + 2]) << 16
             word |= UInt32(bytes[i + 3]) << 24
             
-            hash = update(hash, word)
+            hash = update2(hash, Int(word))
         }
         let remaining = byteCount & 3
         if remaining != 0 {
